requirements.txt
boto3
opencv-python
python-dotenv
streamlit
fastapi
uvicorn